feature #237: make tooltip an atom
